Fix dimension mismatch in texture write operations by ensuring bytes per pixel is cast to usize. Updated expected data length calculations in `record_commands_to_buffer` for consistency.

+++ src/backend/metal/compute.rs
@@ -396,7 +396,8 @@ fn record_commands_to_buffer(
                     anyhow::bail!("WriteTexture: dimension mismatch");
                 }
                 let bpp = tex_state.format.bytes_per_pixel();
-                let expected = (*width as usize) * (*height as usize) * bpp;
+                let expected =
+                    (*width as usize) * (*height as usize) * (bpp as usize);
                 if data.len() != expected {
                     end_compute!();
                     anyhow::bail!(
@@ -456,7 +457,8 @@ fn record_commands_to_buffer(
                     anyhow::bail!("WriteTextureRegion: region out of bounds");
                 }
                 let bpp = tex_state.format.bytes_per_pixel();
-                let expected = (*width as usize) * (*height as usize) * bpp;
+                let expected =
+                    (*width as usize) * (*height as usize) * (bpp as usize);
                 if data.len() != expected {
                     end_compute!();
                     anyhow::bail!(
